P0/back/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy